+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+
+## [0.2.0] - 2025-11-19
+
+### Added
+- Eleveld 2018 PK/PD model
+- GitHub workflows for linting, unit tests and large files
+
 ## [0.1.0] - 2025-11-19
 
 ### Added